Move completion priority (value/similarity) to a common utility function
